requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
PyJWT==2.5.0
pytz==2022.2.1
PyYAML==6.0
sqlparse==0.4.3
tzdata==2022.2
uritemplate==4.1.1
coverage==6.5.0
